+++ src/Cropperjs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Minimum required Symfony version is now 6.4
 - Minimum required PHP version is now 8.2
+- Always apply rotation in `Crop::getCroppedImage()` and `Crop::getCroppedThumbnail()`
 
 ## 2.34